Move connectors under the database umbrella
