+++ database/connectors/janusgraph/things_helper.go
