+++ database/connectors/janusgraph/token.go
@@ -11,7 +11,7 @@ import (
 	"github.com/creativesoftwarefdn/weaviate/gremlin"
 	"github.com/creativesoftwarefdn/weaviate/models"
 
-	connutils "github.com/creativesoftwarefdn/weaviate/connectors/utils"
+	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
 )
 
 // ValidateToken validates/gets a key to the Grelmin database with the given token (=UUID)
