+++ database/connectors/utils/helper_functions.go
