+++ database/connectors/utils/helper_objects.go
