+++ restapi/configure_weaviate.go
@@ -44,10 +44,10 @@ import (
 	"github.com/creativesoftwarefdn/weaviate/auth"
 	"github.com/creativesoftwarefdn/weaviate/broker"
 	"github.com/creativesoftwarefdn/weaviate/config"
-	dbconnector "github.com/creativesoftwarefdn/weaviate/connectors"
-	dblisting "github.com/creativesoftwarefdn/weaviate/connectors/listing"
-	"github.com/creativesoftwarefdn/weaviate/connectors/utils"
 	"github.com/creativesoftwarefdn/weaviate/database"
+	dbconnector "github.com/creativesoftwarefdn/weaviate/database/connectors"
+	dblisting "github.com/creativesoftwarefdn/weaviate/database/connectors/listing"
+	"github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
 	db_schema "github.com/creativesoftwarefdn/weaviate/database/schema"
 	db_local_schema_manager "github.com/creativesoftwarefdn/weaviate/database/schema_manager/local"
 	"github.com/creativesoftwarefdn/weaviate/graphqlapi"
