+++ test/acceptance/things/things_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/creativesoftwarefdn/weaviate/test/acceptance/helper"
 	"github.com/creativesoftwarefdn/weaviate/validation"
 
-	connutils "github.com/creativesoftwarefdn/weaviate/connectors/utils"
+	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
 )
 
 const fakeThingId strfmt.UUID = "11111111-1111-1111-1111-111111111111"
